Get landing page css working
