+++ client/components/SwitchContainer.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import LandingPage from './LandingPage';
-import HomeContainer from './home/HomeContainer';
+import LandingPage from './pages/LandingPage';
+import HomePage from './pages/HomePage';
 
 class SwitchContainer extends React.Component {
 	render() {
-		return this.props.authenticated ? <HomeContainer /> : <LandingPage />;
+		return this.props.authenticated ? <HomePage /> : <LandingPage />;
 	}
 }
